Remove async traits and rework all kernels  (#34)
* Dont always wait for completion

* Remove some async fns

* Remove some async trait

* Remove some tokio::test

* Rework all kernels and introduce kernels which take pipeline as parameter

+++ crates/logical/compute_shaders/i16/shift.wgsl
@@ -28,11 +28,12 @@ fn bitwise_shr(@builtin(global_invocation_id) global_id: vec3<u32>) {
     new_values[global_id.x] = (left_u16 & MAX_I16) | ((right_u16 & MAX_I16) << 16u) ;
 }
 
+// i32(0x00008000) is done temporarily until https://github.com/gfx-rs/wgpu/issues/5085 is resolved
 fn shr(input: i32, shift_value: u32) -> i32 {
     if input < 0 {
         var result = input >> shift_value;
         var other_bit_sets = MAX_I16 << (16u - shift_value);
-        other_bit_sets = other_bit_sets | (0x00008000 >> shift_value);
+        other_bit_sets = other_bit_sets | (i32(0x00008000) >> shift_value);
         return other_bit_sets | result;
     } else {
         return input >> shift_value;
